+++ magpie/api/management/user/user_utils.py
@@ -310,7 +310,7 @@ def create_user_resource_permission_response(user, resource, permission, db_sess
     ax.evaluate_call(lambda: db_session.add(new_perm), fallback=lambda: db_session.rollback(),
                      http_error=HTTPForbidden, content=err_content,
                      msg_on_fail=s.UserResourcePermissions_POST_ForbiddenResponseSchema.description)
-    webhook_params = get_permission_update_params(user, resource, permission)
+    webhook_params = get_permission_update_params(user, resource, permission, db_session)
     process_webhook_requests(WebhookAction.CREATE_USER_PERMISSION, webhook_params)
     return ax.valid_http(http_success=http_success, content=err_content, detail=http_detail)
 
@@ -470,7 +470,7 @@ def delete_user_resource_permission_response(user, resource, permission, db_sess
     ax.evaluate_call(lambda: db_session.delete(del_perm), fallback=lambda: db_session.rollback(),
                      http_error=HTTPNotFound, content=err_content,
                      msg_on_fail=s.UserResourcePermissionName_DELETE_NotFoundResponseSchema.description)
-    webhook_params = get_permission_update_params(user, resource, permission)
+    webhook_params = get_permission_update_params(user, resource, permission, db_session)
     process_webhook_requests(WebhookAction.DELETE_USER_PERMISSION, webhook_params)
     return ax.valid_http(http_success=HTTPOk, detail=s.UserResourcePermissionName_DELETE_OkResponseSchema.description)